requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.9.0
django-filter==2.0.0
djangorestframework==3.16.1
mock==2.0.0
pbr==7.0.1
Pillow==2.1.0
pytz==2025.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
